Added player firing
Explodes on reaching the top of the play area.
Locking mechanism will prevent another shot being fired until the fire
button is released after the shot explodes so the player can't just keep
fire pressed all the time (this becomes more meaningful when the invader
get closer to the base).